+++ src/main/java/pixelitor/filters/levels/GrayScaleLookup.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Laszlo Balazs-Csiki and Contributors
+ * Copyright 2020 Laszlo Balazs-Csiki and Contributors
  *
  * This file is part of Pixelitor. Pixelitor is free software: you
  * can redistribute it and/or modify it under the terms of the GNU
@@ -38,7 +38,7 @@ public class GrayScaleLookup {
             multiplier = 0;
             constant = 255 - inputDarkValue;
         } else {
-            multiplier = (double) (outputLightValue - outputDarkValue)
+            multiplier = (outputLightValue - outputDarkValue)
                     / (double) inputDiff;
             constant = outputDarkValue - multiplier * inputDarkValue;
         }
